Create README.md
I am new to the field entirely and am learning the basics from the ground up.  I intend to go to a coding boot camp once I feel comfortable in the understanding of the basic principals of coding and GPIO.  Any help, references to learning material or suggestions on the subject have my upmost gratitude.